Move load function to separate file.
