+++ ytree/data_structures/__init__.py
@@ -13,12 +13,5 @@ arbor imports
 # The full license is in the file COPYING.txt, distributed with this software.
 #-----------------------------------------------------------------------------
 
-from ytree.data_structures.arbor import \
-    Arbor, \
-    load
-from ytree.data_structures.tree_node_selector import \
-    TreeNodeSelector, \
-    add_tree_node_selector
-
 from ytree.frontends.api import _frontend_container
 frontends = _frontend_container()
